requirements.txt
pandas
numpy
nltk
textblob
torch==1.8.1
transformers==4.2.1
matplotlib
pyyaml